requirements.txt
# Core A2A Server & Client
fastapi==0.133.1
uvicorn==0.41.0
httpx==0.28.1
httpx-sse==0.4.3
starlette==0.52.1
python-multipart==0.0.22
sse-starlette==3.2.0

# UI
gradio==6.6.0
gradio-client==2.1.0

# Data Validation
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
annotated-types==0.7.0

# Google Gemini (used by Orchestrator)
google-genai==1.65.0

# Google ADK (used by org_a_adk variant)
google-adk==1.25.1
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.139.0
google-cloud-appengine-logging==1.8.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.40.1
google-cloud-bigquery-storage==2.36.2
google-cloud-bigtable==2.35.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.21.0
google-cloud-logging==3.13.0
google-cloud-monitoring==2.29.1
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-secret-manager==2.26.0
google-cloud-spanner==3.63.0
google-cloud-speech==2.36.1
google-cloud-storage==3.9.0
google-cloud-trace==1.18.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.1
grpcio-status==1.78.1
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.1

# LangGraph (used by org_b_langgraph variant)
langgraph==1.0.9
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langchain-core==1.2.16
langchain-google-genai==4.2.1
langsmith==0.7.7

# Environment & Config
python-dotenv==1.2.1

# HTTP & Networking
httpcore==1.0.9
httplib2==0.31.2
urllib3==2.6.3
requests==2.32.5
requests-toolbelt==1.0.0
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11

# Async & Runtime
anyio==4.12.1
sniffio==1.3.1
h11==0.16.0
aiofiles==24.1.0
aiosqlite==0.22.1
websockets==15.0.1
watchdog==6.0.0

# Serialization & Schema
orjson==3.11.7
ormsgpack==1.12.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
pyyaml==6.0.3

# CLI & Formatting
typer==0.24.1
typer-slim==0.24.0
click==8.3.1
rich==14.3.3
pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
shellingham==1.5.4

# MCP (Model Context Protocol - used by ADK)
mcp==1.26.0

# Auth & Crypto
authlib==1.6.8
cryptography==46.0.5
pyopenssl==25.3.0
pyjwt==2.11.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
rsa==4.9.1
cffi==2.0.0
pycparser==3.0

# OpenTelemetry (used by ADK internals)
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-resourcedetector-gcp==1.11.0a0

# Data & ML support
numpy==2.4.2
pandas==3.0.1
pillow==12.1.1

# Hugging Face (used by Gradio)
huggingface-hub==1.4.1
safehttpx==0.1.7
hf-xet==1.3.1
fsspec==2026.2.0
filelock==3.24.3

# DB & Migrations (used by ADK session/memory services)
sqlalchemy==2.0.47
sqlalchemy-spanner==1.17.2
alembic==1.18.4
mako==1.3.10

# Utilities
packaging==26.0
typing-extensions==4.15.0
typing-inspection==0.4.2
importlib-metadata==8.7.1
zipp==3.23.0
six==1.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
tzlocal==5.3.1
pyparsing==3.3.2
distro==1.9.0
docstring-parser==0.17.0
tenacity==9.1.4
tqdm==4.67.3
cloudpickle==3.1.2
graphviz==0.21
mmh3==5.2.0
xxhash==3.6.0
uuid-utils==0.14.1
uritemplate==4.2.0
attrs==25.4.0
semantic-version==2.10.0
tomlkit==0.13.3
filetype==1.2.0
ffmpy==1.0.0
pydub==0.25.1
brotli==1.2.0
grpc-interceptor==0.15.4
markupsafe==3.0.3
jinja2==3.1.6
sqlparse==0.5.5
zstandard==0.25.0
audioop-lts==0.2.2
